Addressing review comments  - include cleanup  - enum to enum class  - catch2 as submodule  - avoid usage of c cast  - select serial or parallel implementation depending on MPI and HDF5 found
Fixed bug when writing to HDF5 file on serial implementation

+++ src/reports/library/compartment_report.hpp
@@ -5,6 +5,6 @@ class CompartmentReport: public Report {
   public:
     CompartmentReport(const std::string& reportName, double tstart, double tend, double dt);
 
-    int get_total_compartments() override;
+    size_t get_total_compartments() override;
     int add_variable(int cell_number, double* pointer) override;
 };
